fix the types and remove the dependency to custom types

+++ src/main.cpp
@@ -5,7 +5,7 @@
 #include "sort.hpp"
 
 int main() {
-    uint arr[] = {3, 8, 5, 4, 100};
+    int64_t arr[] = {3, 8, 5, 4, 100};
     std::cout << "Before:" << std::endl;
     print_arr(arr, 5, std::string("arr"));
 
@@ -13,7 +13,7 @@ int main() {
     std::cout << std::endl << "After:" << std::endl;
     print_arr(arr, 5, std::string("arr"));
 
-    uint arr2[] = {3, 8, 5, 4, 100, 50};
+    int64_t arr2[] = {3, 8, 5, 4, 100, 50};
     std::cout << "Before:" << std::endl;
     print_arr(arr2, 6, std::string("arr"));
