+++ src/printer.cpp
@@ -2,10 +2,10 @@
 
 #include "printer.hpp"
 
-void print_arr(uint *arr, usize n, std::string name) {
+void print_arr(int64_t *arr, uintptr_t n, std::string name) {
     std::cout << name << " = { ";
 
-    for (usize i = 0; i < n; ++i) {
+    for (uintptr_t i = 0; i < n; ++i) {
         std::cout << arr[i];
 
         if (i != n - 1) {
